:recycle: align jest test structure with src folder
