+++ packages/core/tests/unit/renderer.test.js
@@ -12,140 +12,137 @@ const Printer = require("@graphql-markdown/printer-legacy");
 const Renderer = require("../../src/renderer");
 const { GraphQLObjectType } = require("graphql");
 
-describe("lib", () => {
-  describe("renderer", () => {
-    describe("class Renderer", () => {
-      let rendererInstance;
-      let baseURL = "graphql";
-      let printerInstance;
+describe("renderer", () => {
+  describe("class Renderer", () => {
+    let rendererInstance;
+    let baseURL = "graphql";
+    let printerInstance;
 
-      beforeEach(() => {
-        jest.resetModules();
+    beforeEach(() => {
+      jest.resetModules();
 
-        vol.fromJSON({
-          "/output": {},
-          "/temp": {},
-          "/assets/generated.md": "Test Homepage",
-        });
-
-        printerInstance = new Printer("SCHEMA", baseURL, "root");
-        rendererInstance = new Renderer(printerInstance, "/output", baseURL);
+      vol.fromJSON({
+        "/output": {},
+        "/temp": {},
+        "/assets/generated.md": "Test Homepage",
       });
 
-      afterEach(() => {
-        vol.reset();
-      });
+      printerInstance = new Printer("SCHEMA", baseURL, "root");
+      rendererInstance = new Renderer(printerInstance, "/output", baseURL);
+    });
 
-      describe("renderTypeEntities()", () => {
-        test("creates entity page into output folder", async () => {
-          expect.assertions(2);
+    afterEach(() => {
+      vol.reset();
+    });
 
-          jest
-            .spyOn(printerInstance, "printType")
-            .mockReturnValue("Lorem ipsum");
-          const output = "/output/foobar";
+    describe("renderTypeEntities()", () => {
+      test("creates entity page into output folder", async () => {
+        expect.assertions(2);
 
-          const meta = await rendererInstance.renderTypeEntities(
-            output,
-            "FooBar",
-            "FooBar",
-          );
+        jest.spyOn(printerInstance, "printType").mockReturnValue("Lorem ipsum");
+        const output = "/output/foobar";
 
-          expect(meta).toEqual({ category: "Foobar", slug: "foobar/foo-bar" });
-          expect(vol.toJSON("/output", undefined, true)).toMatchSnapshot();
-        });
-
-        test.each([[undefined, null]])(
-          "do nothing if type is not defined",
-          async (type) => {
-            expect.assertions(1);
-            const meta = await rendererInstance.renderTypeEntities(
-              "test",
-              "FooBar",
-              type,
-            );
-            expect(meta).toBeUndefined();
-          },
+        const meta = await rendererInstance.renderTypeEntities(
+          output,
+          "FooBar",
+          "FooBar",
         );
+
+        expect(meta).toEqual({ category: "Foobar", slug: "foobar/foo-bar" });
+        expect(vol.toJSON("/output", undefined, true)).toMatchSnapshot();
       });
 
-      describe("renderSidebar()", () => {
-        test("creates Docusaurus compatible sidebar.js into output folder", async () => {
+      test.each([[undefined, null]])(
+        "do nothing if type is not defined",
+        async (type) => {
           expect.assertions(1);
-
-          await rendererInstance.renderSidebar();
-
-          expect(
-            vol.toJSON("/output/sidebar-schema.js", undefined, true),
-          ).toMatchSnapshot();
-        });
-      });
-
-      describe("renderHomepage()", () => {
-        test("copies default homepage into output folder", async () => {
-          expect.assertions(1);
-
-          await rendererInstance.renderHomepage("/assets/generated.md");
-
-          expect(vol.toJSON("/output", undefined, true)).toMatchSnapshot();
-        });
-      });
-
-      describe("renderRootTypes()", () => {
-        test("render root type", async () => {
-          expect.assertions(1);
-
-          jest
-            .spyOn(printerInstance, "printType")
-            .mockImplementation(() => "content");
-          await rendererInstance.renderRootTypes("Object", {
-            foo: new GraphQLObjectType({ name: "foo", astNode: {} }),
-            bar: new GraphQLObjectType({ name: "bar", astNode: {} }),
-          });
-
-          expect(vol.toJSON("/output", undefined, true)).toMatchSnapshot();
-        });
-      });
-
-      describe("generateCategoryMetafile()", () => {
-        test("generate _category_.yml file", async () => {
-          expect.assertions(2);
-
-          const category = "foobar";
-          const outputPath = "/output/docs";
-
-          await rendererInstance.generateCategoryMetafile(category, outputPath);
-
-          const content = fs.readFileSync(
-            path.join(outputPath, "_category_.yml"),
-            "utf-8",
+          const meta = await rendererInstance.renderTypeEntities(
+            "test",
+            "FooBar",
+            type,
           );
+          expect(meta).toBeUndefined();
+        },
+      );
+    });
 
-          expect(vol.toJSON("/output", undefined, true)).toMatchSnapshot();
-          expect(content).toMatchInlineSnapshot(`
+    describe("renderSidebar()", () => {
+      test("creates Docusaurus compatible sidebar.js into output folder", async () => {
+        expect.assertions(1);
+
+        await rendererInstance.renderSidebar();
+
+        expect(
+          vol.toJSON("/output/sidebar-schema.js", undefined, true),
+        ).toMatchSnapshot();
+      });
+    });
+
+    describe("renderHomepage()", () => {
+      test("copies default homepage into output folder", async () => {
+        expect.assertions(1);
+
+        await rendererInstance.renderHomepage("/assets/generated.md");
+
+        expect(vol.toJSON("/output", undefined, true)).toMatchSnapshot();
+      });
+    });
+
+    describe("renderRootTypes()", () => {
+      test("render root type", async () => {
+        expect.assertions(1);
+
+        jest
+          .spyOn(printerInstance, "printType")
+          .mockImplementation(() => "content");
+        await rendererInstance.renderRootTypes("Object", {
+          foo: new GraphQLObjectType({ name: "foo", astNode: {} }),
+          bar: new GraphQLObjectType({ name: "bar", astNode: {} }),
+        });
+
+        expect(vol.toJSON("/output", undefined, true)).toMatchSnapshot();
+      });
+    });
+
+    describe("generateCategoryMetafile()", () => {
+      test("generate _category_.yml file", async () => {
+        expect.assertions(2);
+
+        const category = "foobar";
+        const outputPath = "/output/docs";
+
+        await rendererInstance.generateCategoryMetafile(category, outputPath);
+
+        const content = fs.readFileSync(
+          path.join(outputPath, "_category_.yml"),
+          "utf-8",
+        );
+
+        expect(vol.toJSON("/output", undefined, true)).toMatchSnapshot();
+        expect(content).toMatchInlineSnapshot(`
             "label: Foobar
             link: null
             "
           `);
-        });
+      });
 
-        test("generate _category_.yml file with generated index", async () => {
-          expect.assertions(2);
+      test("generate _category_.yml file with generated index", async () => {
+        expect.assertions(2);
 
-          const category = "foobar";
-          const outputPath = "/output/docs";
+        const category = "foobar";
+        const outputPath = "/output/docs";
 
-          rendererInstance.options = { index: true };
+        rendererInstance.options = { index: true };
 
-          await rendererInstance.generateCategoryMetafile(category, outputPath);
+        await rendererInstance.generateCategoryMetafile(category, outputPath);
 
-          const content = fs.readFileSync(
-            path.join(outputPath, "_category_.yml"),
-            "utf-8",
-          );
+        const content = fs.readFileSync(
+          path.join(outputPath, "_category_.yml"),
+          "utf-8",
+        );
 
-          expect(vol.toJSON("/output", undefined, true)).toMatchSnapshot();
-          expect(content).toMatchInlineSnapshot(`
+        expect(vol.toJSON("/output", undefined, true)).toMatchSnapshot();
+        expect(content).toMatchInlineSnapshot(`
             "label: Foobar
             link: 
               type: generated-index
@@ -153,32 +150,31 @@ describe("lib", () => {
 
             "
           `);
-        });
+      });
 
-        test("do not generate _category_.yml file if it exists", async () => {
-          expect.assertions(1);
+      test("do not generate _category_.yml file if it exists", async () => {
+        expect.assertions(1);
 
-          const category = "foobar";
-          const outputPath = "/output/docs";
+        const category = "foobar";
+        const outputPath = "/output/docs";
 
-          const data = "The quick brown fox jumps over the lazy dog";
+        const data = "The quick brown fox jumps over the lazy dog";
 
-          await ensureDir(outputPath);
-          fs.writeFileSync(
-            path.join(outputPath, "_category_.yml"),
-            data,
-            "utf-8",
-          );
+        await ensureDir(outputPath);
+        fs.writeFileSync(
+          path.join(outputPath, "_category_.yml"),
+          data,
+          "utf-8",
+        );
 
-          await rendererInstance.generateCategoryMetafile(category, outputPath);
+        await rendererInstance.generateCategoryMetafile(category, outputPath);
 
-          const content = fs.readFileSync(
-            path.join(outputPath, "_category_.yml"),
-            "utf-8",
-          );
+        const content = fs.readFileSync(
+          path.join(outputPath, "_category_.yml"),
+          "utf-8",
+        );
 
-          expect(content).toEqual(data);
-        });
+        expect(content).toEqual(data);
       });
     });
   });
