+++ packages/utils/tests/unit/fs.test.js
@@ -3,81 +3,79 @@ jest.mock("fs");
 
 const { ensureDir, fileExists, saveFile } = require("../../src/fs");
 
-describe("helpers", () => {
-  describe("fs", () => {
-    beforeEach(() => {
-      vol.fromJSON({
-        "/testFolder": {},
-        "/testFolder/testFile": "just a test",
-      });
-    });
-
-    afterEach(() => {
-      vol.reset();
-    });
-
-    describe("fileExists()", () => {
-      test.each([
-        {
-          type: "file",
-          path: "/testFolder/testFile",
-          expected: true,
-          desc: "exists",
-        },
-        { type: "folder", path: "/testFolder", expected: true, desc: "exists" },
-        {
-          type: "file",
-          path: "/testFolder/missingFile",
-          expected: false,
-          desc: "not exists",
-        },
-        {
-          type: "folder",
-          path: "/missingFolder",
-          expected: false,
-          desc: "not exists",
-        },
-      ])(
-        "return $expected if $type '$path' $desc",
-        async ({ path, expected }) => {
-          expect.assertions(1);
-
-          await expect(fileExists(path)).resolves.toBe(expected);
-        },
-      );
-    });
-
-    describe("ensureDir()", () => {
-      test.each([
-        { path: "/testFolder", exists: true, desc: "not created if present" },
-        {
-          path: "/testNewFolder",
-          exists: false,
-          desc: "created if not present",
-        },
-      ])("folder is", async ({ path, exists }) => {
-        expect.assertions(2);
-
-        await expect(fileExists(path)).resolves.toBe(exists);
-
-        await ensureDir(path);
-
-        await expect(fileExists(path)).resolves.toBeTruthy();
-      });
+describe("fs", () => {
+  beforeEach(() => {
+    vol.fromJSON({
+      "/testFolder": {},
+      "/testFolder/testFile": "just a test",
     });
   });
 
-  describe("saveFile()", () => {
-    test("create file and folders", async () => {
-      expect.assertions(1);
+  afterEach(() => {
+    vol.reset();
+  });
 
-      await saveFile("/foo/bar/test/foobar.test", "foobar file for test");
+  describe("fileExists()", () => {
+    test.each([
+      {
+        type: "file",
+        path: "/testFolder/testFile",
+        expected: true,
+        desc: "exists",
+      },
+      { type: "folder", path: "/testFolder", expected: true, desc: "exists" },
+      {
+        type: "file",
+        path: "/testFolder/missingFile",
+        expected: false,
+        desc: "not exists",
+      },
+      {
+        type: "folder",
+        path: "/missingFolder",
+        expected: false,
+        desc: "not exists",
+      },
+    ])(
+      "return $expected if $type '$path' $desc",
+      async ({ path, expected }) => {
+        expect.assertions(1);
 
-      expect(vol.toJSON("/foo/bar/test/foobar.test")).toMatchInlineSnapshot(`
+        await expect(fileExists(path)).resolves.toBe(expected);
+      },
+    );
+  });
+
+  describe("ensureDir()", () => {
+    test.each([
+      { path: "/testFolder", exists: true, desc: "not created if present" },
+      {
+        path: "/testNewFolder",
+        exists: false,
+        desc: "created if not present",
+      },
+    ])("folder is", async ({ path, exists }) => {
+      expect.assertions(2);
+
+      await expect(fileExists(path)).resolves.toBe(exists);
+
+      await ensureDir(path);
+
+      await expect(fileExists(path)).resolves.toBeTruthy();
+    });
+  });
+});
+
+describe("saveFile()", () => {
+  test("create file and folders", async () => {
+    expect.assertions(1);
+
+    await saveFile("/foo/bar/test/foobar.test", "foobar file for test");
+
+    expect(vol.toJSON("/foo/bar/test/foobar.test")).toMatchInlineSnapshot(`
         {
           "/foo/bar/test/foobar.test": "foobar file for test",
         }
       `);
-    });
   });
 });
