+++ packages/utils/tests/unit/graphql.test.js
@@ -39,423 +39,422 @@ const SCHEMA_CUSTOM_ROOT_FILE = require.resolve(
   "../__data__/schema_with_custom_root_types.graphql",
 );
 
-describe("lib", () => {
-  describe("graphql", () => {
-    let schema;
+describe("graphql", () => {
+  let schema;
 
-    beforeAll(async () => {
-      schema = await gqlToolsLoadSchema(SCHEMA_FILE, {
+  beforeAll(async () => {
+    schema = await gqlToolsLoadSchema(SCHEMA_FILE, {
+      loaders: [new GraphQLFileLoader()],
+    });
+  });
+
+  describe("loadSchema()", () => {
+    test("returns valid schema", async () => {
+      const testSchema = await loadSchema(SCHEMA_FILE, {
         loaders: [new GraphQLFileLoader()],
       });
+      expect(JSON.stringify(testSchema)).toBe(JSON.stringify(schema));
     });
 
-    describe("loadSchema()", () => {
-      test("returns valid schema", async () => {
-        const testSchema = await loadSchema(SCHEMA_FILE, {
-          loaders: [new GraphQLFileLoader()],
-        });
-        expect(JSON.stringify(testSchema)).toBe(JSON.stringify(schema));
+    test("returns valid schema with custom root type", async () => {
+      const testSchema = await loadSchema(SCHEMA_CUSTOM_ROOT_FILE, {
+        loaders: [new GraphQLFileLoader()],
+        rootTypes: { query: "Root", subscription: "" },
       });
 
-      test("returns valid schema with custom root type", async () => {
-        const testSchema = await loadSchema(SCHEMA_CUSTOM_ROOT_FILE, {
-          loaders: [new GraphQLFileLoader()],
-          rootTypes: { query: "Root", subscription: "" },
-        });
+      expect(testSchema.getQueryType().name).toBe("Root");
+      expect(testSchema.getMutationType()).toBeUndefined();
+      expect(testSchema.getSubscriptionType()).toBeUndefined();
+    });
+  });
 
-        expect(testSchema.getQueryType().name).toBe("Root");
-        expect(testSchema.getMutationType()).toBeUndefined();
-        expect(testSchema.getSubscriptionType()).toBeUndefined();
-      });
+  describe("getDocumentLoaders()", () => {
+    test("returns loaders when plugin config loaders format is a string", () => {
+      const loaders = {
+        GraphQLFileLoader: "@graphql-tools/graphql-file-loader",
+      };
+      const { loaders: documentLoaders, loaderOptions } =
+        getDocumentLoaders(loaders);
+
+      expect(documentLoaders).toMatchObject([new GraphQLFileLoader()]);
+      expect(loaderOptions).toMatchObject({});
     });
 
-    describe("getDocumentLoaders()", () => {
-      test("returns loaders when plugin config loaders format is a string", () => {
-        const loaders = {
-          GraphQLFileLoader: "@graphql-tools/graphql-file-loader",
-        };
-        const { loaders: documentLoaders, loaderOptions } =
-          getDocumentLoaders(loaders);
-
-        expect(documentLoaders).toMatchObject([new GraphQLFileLoader()]);
-        expect(loaderOptions).toMatchObject({});
-      });
-
-      test("returns loaders and configuration when plugin config loaders format is an object", () => {
-        const loaders = {
-          GraphQLFileLoader: {
-            module: "@graphql-tools/graphql-file-loader",
-            options: {
-              option1: true,
-            },
+    test("returns loaders and configuration when plugin config loaders format is an object", () => {
+      const loaders = {
+        GraphQLFileLoader: {
+          module: "@graphql-tools/graphql-file-loader",
+          options: {
+            option1: true,
           },
-        };
-        const { loaders: documentLoaders, loaderOptions } =
-          getDocumentLoaders(loaders);
+        },
+      };
+      const { loaders: documentLoaders, loaderOptions } =
+        getDocumentLoaders(loaders);
 
-        expect(documentLoaders).toMatchObject([new GraphQLFileLoader()]);
-        expect(loaderOptions).toMatchObject({
-          option1: true,
-        });
+      expect(documentLoaders).toMatchObject([new GraphQLFileLoader()]);
+      expect(loaderOptions).toMatchObject({
+        option1: true,
       });
     });
+  });
 
-    // covers printDefaultValue()
-    describe("getDefaultValue()", () => {
-      test.each([
-        { type: GraphQLInt, value: 5 },
-        { type: GraphQLInt, value: 0 },
-        { type: GraphQLFloat, value: 5.3 },
-        { type: GraphQLFloat, value: 0.0 },
-        { type: GraphQLBoolean, value: true },
-        { type: GraphQLBoolean, value: false },
-      ])("returns $value value as default for $type", ({ type, value }) => {
+  // covers printDefaultValue()
+  describe("getDefaultValue()", () => {
+    test.each([
+      { type: GraphQLInt, value: 5 },
+      { type: GraphQLInt, value: 0 },
+      { type: GraphQLFloat, value: 5.3 },
+      { type: GraphQLFloat, value: 0.0 },
+      { type: GraphQLBoolean, value: true },
+      { type: GraphQLBoolean, value: false },
+    ])("returns $value value as default for $type", ({ type, value }) => {
+      expect.hasAssertions();
+
+      const argument = {
+        name: "foobar",
+        description: undefined,
+        type: type,
+        defaultValue: value,
+        extensions: undefined,
+      };
+
+      expect(getDefaultValue(argument)).toEqual(value);
+    });
+
+    test.each([
+      { type: GraphQLInt },
+      { type: GraphQLID },
+      { type: GraphQLFloat },
+      { type: GraphQLString },
+      { type: GraphQLBoolean },
+      { type: new GraphQLList(GraphQLID) },
+    ])(
+      "returns undefined for type $type if not default value defined",
+      ({ type }) => {
         expect.hasAssertions();
 
         const argument = {
           name: "foobar",
           description: undefined,
           type: type,
-          defaultValue: value,
+          defaultValue: undefined,
           extensions: undefined,
         };
 
-        expect(getDefaultValue(argument)).toEqual(value);
-      });
+        expect(getDefaultValue(argument)).toBeUndefined();
+      },
+    );
 
-      test.each([
-        { type: GraphQLInt },
-        { type: GraphQLID },
-        { type: GraphQLFloat },
-        { type: GraphQLString },
-        { type: GraphQLBoolean },
-        { type: new GraphQLList(GraphQLID) },
-      ])(
-        "returns undefined for type $type if not default value defined",
-        ({ type }) => {
-          expect.hasAssertions();
-
-          const argument = {
-            name: "foobar",
-            description: undefined,
-            type: type,
-            defaultValue: undefined,
-            extensions: undefined,
-          };
-
-          expect(getDefaultValue(argument)).toBeUndefined();
-        },
-      );
-
-      test.each([
-        {
-          type: new GraphQLList(GraphQLID),
-          defaultValue: ["0", "1"],
-          expected: '["0", "1"]',
-        },
-        {
-          type: new GraphQLList(GraphQLInt),
-          defaultValue: 42,
-          expected: "[42]",
-        },
-      ])(
-        "returns array default value as string for type $type",
-        ({ type, defaultValue, expected }) => {
-          expect.hasAssertions();
-
-          const argument = {
-            name: "id",
-            description: undefined,
-            type,
-            defaultValue,
-            extensions: undefined,
-          };
-
-          expect(getDefaultValue(argument)).toBe(expected);
-        },
-      );
-
-      test("returns unformatted default value for type GraphQLEnum", () => {
+    test.each([
+      {
+        type: new GraphQLList(GraphQLID),
+        defaultValue: ["0", "1"],
+        expected: '["0", "1"]',
+      },
+      {
+        type: new GraphQLList(GraphQLInt),
+        defaultValue: 42,
+        expected: "[42]",
+      },
+    ])(
+      "returns array default value as string for type $type",
+      ({ type, defaultValue, expected }) => {
         expect.hasAssertions();
-
-        const enumType = new GraphQLEnumType({
-          name: "RGB",
-          values: {
-            RED: { value: "RED" },
-            GREEN: { value: "GREEN" },
-            BLUE: { value: "BLUE" },
-          },
-        });
 
         const argument = {
-          name: "color",
+          name: "id",
           description: undefined,
-          type: enumType,
-          defaultValue: "RED",
+          type,
+          defaultValue,
           extensions: undefined,
         };
 
-        expect(getDefaultValue(argument)).toBe("RED");
-      });
+        expect(getDefaultValue(argument)).toBe(expected);
+      },
+    );
 
-      test("returns array default value unformatted for type GraphQLList(GraphQLEnum)", () => {
-        expect.hasAssertions();
+    test("returns unformatted default value for type GraphQLEnum", () => {
+      expect.hasAssertions();
 
-        const enumType = new GraphQLEnumType({
-          name: "RGB",
-          values: {
-            RED: { value: "RED" },
-            GREEN: { value: "GREEN" },
-            BLUE: { value: "BLUE" },
-          },
-        });
-
-        const argument = {
-          name: "color",
-          description: undefined,
-          type: new GraphQLList(enumType),
-          defaultValue: ["RED"],
-          extensions: undefined,
-        };
-
-        expect(getDefaultValue(argument)).toBe("[RED]");
-      });
-    });
-
-    describe("getIntrospectionFieldsList()", () => {
-      test("returns list of queries", () => {
-        expect.hasAssertions();
-
-        const list = getIntrospectionFieldsList(schema.getQueryType());
-
-        expect(JSON.stringify(list, null, 2)).toMatchSnapshot();
-      });
-
-      test("returns list of mutations", () => {
-        expect.hasAssertions();
-
-        const list = getIntrospectionFieldsList(schema.getMutationType());
-
-        expect(JSON.stringify(list, null, 2)).toMatchSnapshot();
-      });
-
-      test("returns list of subscriptions", () => {
-        expect.hasAssertions();
-
-        const list = getIntrospectionFieldsList(schema.getSubscriptionType());
-
-        expect(JSON.stringify(list, null, 2)).toMatchSnapshot();
-      });
-
-      test("returns undefined if null", () => {
-        expect.hasAssertions();
-
-        const list = getIntrospectionFieldsList(null);
-
-        expect(list).toBeUndefined();
-      });
-    });
-
-    describe("getFields()", () => {
-      test("returns list of type fields", () => {
-        expect.hasAssertions();
-
-        const fields = getFields(schema.getMutationType());
-
-        expect(JSON.stringify(fields, null, 2)).toMatchSnapshot();
-      });
-
-      test("returns empty list if getFields not supported", () => {
-        expect.hasAssertions();
-
-        const fields = getFields("test");
-
-        expect(fields).toStrictEqual([]);
-      });
-    });
-
-    describe("getTypeName()", () => {
-      test("returns type name for object", () => {
-        expect.hasAssertions();
-
-        const name = getTypeName(schema.getType("Tweet"));
-
-        expect(name).toBe("Tweet");
-      });
-
-      test("returns type name for interface", () => {
-        expect.hasAssertions();
-
-        const name = getTypeName(schema.getType("Node"));
-
-        expect(name).toBe("Node");
-      });
-
-      test("returns type name for scalar", () => {
-        expect.hasAssertions();
-
-        const name = getTypeName(schema.getType("ID"));
-
-        expect(name).toBe("ID");
-      });
-
-      test("returns default name for unknown", () => {
-        expect.hasAssertions();
-
-        const name = getTypeName({ toString: undefined }, "FooBar");
-
-        expect(name).toBe("FooBar");
-      });
-    });
-
-    describe("getTypeFromSchema()", () => {
-      test("returns a filter map filtered by GraphQLObjectType", () => {
-        expect.hasAssertions();
-
-        const map = getTypeFromSchema(schema, GraphQLObjectType);
-
-        expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
-      });
-
-      test("returns a filter map filtered by GraphQLUnionType", () => {
-        expect.hasAssertions();
-
-        const map = getTypeFromSchema(schema, GraphQLUnionType);
-
-        expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
-      });
-
-      test("returns a filter map filtered by GraphQLInterfaceType", () => {
-        expect.hasAssertions();
-
-        const map = getTypeFromSchema(schema, GraphQLInterfaceType);
-
-        expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
-      });
-
-      test("returns a filter map filtered by GraphQLEnumType", () => {
-        expect.hasAssertions();
-
-        const map = getTypeFromSchema(schema, GraphQLEnumType);
-
-        expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
-      });
-
-      test("returns a filter map filtered by GraphQLInputObjectType", () => {
-        expect.hasAssertions();
-
-        const map = getTypeFromSchema(schema, GraphQLInputObjectType);
-
-        expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
-      });
-
-      test("returns a filter map filtered by GraphQLScalarType", () => {
-        expect.hasAssertions();
-
-        const map = getTypeFromSchema(schema, GraphQLScalarType);
-
-        expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
-      });
-
-      test.each([[null], [undefined]])(
-        "returns undefined it typeMap is not defined",
-        (typeMap) => {
-          expect.hasAssertions();
-
-          const map = getTypeFromSchema(typeMap, GraphQLScalarType);
-
-          expect(map).toBeUndefined();
+      const enumType = new GraphQLEnumType({
+        name: "RGB",
+        values: {
+          RED: { value: "RED" },
+          GREEN: { value: "GREEN" },
+          BLUE: { value: "BLUE" },
         },
-      );
+      });
+
+      const argument = {
+        name: "color",
+        description: undefined,
+        type: enumType,
+        defaultValue: "RED",
+        extensions: undefined,
+      };
+
+      expect(getDefaultValue(argument)).toBe("RED");
     });
 
-    describe("getSchemaMap()", () => {
-      test("returns schema types map", () => {
-        expect.hasAssertions();
+    test("returns array default value unformatted for type GraphQLList(GraphQLEnum)", () => {
+      expect.hasAssertions();
 
-        const schemaTypeMap = getSchemaMap(schema);
-
-        expect(JSON.stringify(schemaTypeMap, null, 2)).toMatchSnapshot();
+      const enumType = new GraphQLEnumType({
+        name: "RGB",
+        values: {
+          RED: { value: "RED" },
+          GREEN: { value: "GREEN" },
+          BLUE: { value: "BLUE" },
+        },
       });
 
-      test("returns schema types map with custom root types", async () => {
-        expect.hasAssertions();
+      const argument = {
+        name: "color",
+        description: undefined,
+        type: new GraphQLList(enumType),
+        defaultValue: ["RED"],
+        extensions: undefined,
+      };
 
-        const testSchema = await loadSchema(SCHEMA_CUSTOM_ROOT_FILE, {
-          loaders: [new GraphQLFileLoader()],
-          rootTypes: { query: "Root", subscription: "" },
-        });
-
-        const schemaTypeMap = getSchemaMap(testSchema);
-
-        expect(JSON.stringify(schemaTypeMap, null, 2)).toMatchSnapshot();
-      });
-    });
-
-    describe("isParametrizedField()", () => {
-      test("returns true if type is parametrized", () => {
-        expect.hasAssertions();
-
-        const mutations = getIntrospectionFieldsList(schema.getMutationType());
-        const res = isParametrizedField(mutations["createTweet"]);
-
-        expect(res).toBeTruthy();
-      });
-
-      test("returns false if type is not parametrized", () => {
-        expect.hasAssertions();
-
-        const queries = getIntrospectionFieldsList(schema.getQueryType());
-        const res = isParametrizedField(queries["TweetsMeta"]);
-
-        expect(res).toBeFalsy();
-      });
-    });
-
-    describe("isOperation()", () => {
-      test("returns true if type is mutation", () => {
-        expect.hasAssertions();
-
-        const mutations = getIntrospectionFieldsList(schema.getMutationType());
-        const res = isOperation(mutations["createTweet"]);
-
-        expect(res).toBeTruthy();
-      });
-
-      test("returns true if type is query", () => {
-        expect.hasAssertions();
-
-        const queries = getIntrospectionFieldsList(schema.getQueryType());
-        const res = isOperation(queries["Tweets"]);
-
-        expect(res).toBeTruthy();
-      });
-
-      test("returns true if type is subscription", () => {
-        expect.hasAssertions();
-
-        const subscriptions = getIntrospectionFieldsList(
-          schema.getSubscriptionType(),
-        );
-        const res = isOperation(subscriptions["Notifications"]);
-
-        expect(res).toBeTruthy();
-      });
-
-      test("returns false if type is not an operation", () => {
-        expect.hasAssertions();
-
-        const objects = getTypeFromSchema(schema, GraphQLObjectType);
-        const res = isOperation(objects["Tweet"]);
-
-        expect(res).toBeFalsy();
-      });
+      expect(getDefaultValue(argument)).toBe("[RED]");
     });
   });
 
-  describe("getRelationOfInterface()", () => {
-    test("returns types and interfaces extending an interface", () => {
-      const schema = buildSchema(`
+  describe("getIntrospectionFieldsList()", () => {
+    test("returns list of queries", () => {
+      expect.hasAssertions();
+
+      const list = getIntrospectionFieldsList(schema.getQueryType());
+
+      expect(JSON.stringify(list, null, 2)).toMatchSnapshot();
+    });
+
+    test("returns list of mutations", () => {
+      expect.hasAssertions();
+
+      const list = getIntrospectionFieldsList(schema.getMutationType());
+
+      expect(JSON.stringify(list, null, 2)).toMatchSnapshot();
+    });
+
+    test("returns list of subscriptions", () => {
+      expect.hasAssertions();
+
+      const list = getIntrospectionFieldsList(schema.getSubscriptionType());
+
+      expect(JSON.stringify(list, null, 2)).toMatchSnapshot();
+    });
+
+    test("returns undefined if null", () => {
+      expect.hasAssertions();
+
+      const list = getIntrospectionFieldsList(null);
+
+      expect(list).toBeUndefined();
+    });
+  });
+
+  describe("getFields()", () => {
+    test("returns list of type fields", () => {
+      expect.hasAssertions();
+
+      const fields = getFields(schema.getMutationType());
+
+      expect(JSON.stringify(fields, null, 2)).toMatchSnapshot();
+    });
+
+    test("returns empty list if getFields not supported", () => {
+      expect.hasAssertions();
+
+      const fields = getFields("test");
+
+      expect(fields).toStrictEqual([]);
+    });
+  });
+
+  describe("getTypeName()", () => {
+    test("returns type name for object", () => {
+      expect.hasAssertions();
+
+      const name = getTypeName(schema.getType("Tweet"));
+
+      expect(name).toBe("Tweet");
+    });
+
+    test("returns type name for interface", () => {
+      expect.hasAssertions();
+
+      const name = getTypeName(schema.getType("Node"));
+
+      expect(name).toBe("Node");
+    });
+
+    test("returns type name for scalar", () => {
+      expect.hasAssertions();
+
+      const name = getTypeName(schema.getType("ID"));
+
+      expect(name).toBe("ID");
+    });
+
+    test("returns default name for unknown", () => {
+      expect.hasAssertions();
+
+      const name = getTypeName({ toString: undefined }, "FooBar");
+
+      expect(name).toBe("FooBar");
+    });
+  });
+
+  describe("getTypeFromSchema()", () => {
+    test("returns a filter map filtered by GraphQLObjectType", () => {
+      expect.hasAssertions();
+
+      const map = getTypeFromSchema(schema, GraphQLObjectType);
+
+      expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
+    });
+
+    test("returns a filter map filtered by GraphQLUnionType", () => {
+      expect.hasAssertions();
+
+      const map = getTypeFromSchema(schema, GraphQLUnionType);
+
+      expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
+    });
+
+    test("returns a filter map filtered by GraphQLInterfaceType", () => {
+      expect.hasAssertions();
+
+      const map = getTypeFromSchema(schema, GraphQLInterfaceType);
+
+      expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
+    });
+
+    test("returns a filter map filtered by GraphQLEnumType", () => {
+      expect.hasAssertions();
+
+      const map = getTypeFromSchema(schema, GraphQLEnumType);
+
+      expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
+    });
+
+    test("returns a filter map filtered by GraphQLInputObjectType", () => {
+      expect.hasAssertions();
+
+      const map = getTypeFromSchema(schema, GraphQLInputObjectType);
+
+      expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
+    });
+
+    test("returns a filter map filtered by GraphQLScalarType", () => {
+      expect.hasAssertions();
+
+      const map = getTypeFromSchema(schema, GraphQLScalarType);
+
+      expect(JSON.stringify(map, null, 2)).toMatchSnapshot();
+    });
+
+    test.each([[null], [undefined]])(
+      "returns undefined it typeMap is not defined",
+      (typeMap) => {
+        expect.hasAssertions();
+
+        const map = getTypeFromSchema(typeMap, GraphQLScalarType);
+
+        expect(map).toBeUndefined();
+      },
+    );
+  });
+
+  describe("getSchemaMap()", () => {
+    test("returns schema types map", () => {
+      expect.hasAssertions();
+
+      const schemaTypeMap = getSchemaMap(schema);
+
+      expect(JSON.stringify(schemaTypeMap, null, 2)).toMatchSnapshot();
+    });
+
+    test("returns schema types map with custom root types", async () => {
+      expect.hasAssertions();
+
+      const testSchema = await loadSchema(SCHEMA_CUSTOM_ROOT_FILE, {
+        loaders: [new GraphQLFileLoader()],
+        rootTypes: { query: "Root", subscription: "" },
+      });
+
+      const schemaTypeMap = getSchemaMap(testSchema);
+
+      expect(JSON.stringify(schemaTypeMap, null, 2)).toMatchSnapshot();
+    });
+  });
+
+  describe("isParametrizedField()", () => {
+    test("returns true if type is parametrized", () => {
+      expect.hasAssertions();
+
+      const mutations = getIntrospectionFieldsList(schema.getMutationType());
+      const res = isParametrizedField(mutations["createTweet"]);
+
+      expect(res).toBeTruthy();
+    });
+
+    test("returns false if type is not parametrized", () => {
+      expect.hasAssertions();
+
+      const queries = getIntrospectionFieldsList(schema.getQueryType());
+      const res = isParametrizedField(queries["TweetsMeta"]);
+
+      expect(res).toBeFalsy();
+    });
+  });
+
+  describe("isOperation()", () => {
+    test("returns true if type is mutation", () => {
+      expect.hasAssertions();
+
+      const mutations = getIntrospectionFieldsList(schema.getMutationType());
+      const res = isOperation(mutations["createTweet"]);
+
+      expect(res).toBeTruthy();
+    });
+
+    test("returns true if type is query", () => {
+      expect.hasAssertions();
+
+      const queries = getIntrospectionFieldsList(schema.getQueryType());
+      const res = isOperation(queries["Tweets"]);
+
+      expect(res).toBeTruthy();
+    });
+
+    test("returns true if type is subscription", () => {
+      expect.hasAssertions();
+
+      const subscriptions = getIntrospectionFieldsList(
+        schema.getSubscriptionType(),
+      );
+      const res = isOperation(subscriptions["Notifications"]);
+
+      expect(res).toBeTruthy();
+    });
+
+    test("returns false if type is not an operation", () => {
+      expect.hasAssertions();
+
+      const objects = getTypeFromSchema(schema, GraphQLObjectType);
+      const res = isOperation(objects["Tweet"]);
+
+      expect(res).toBeFalsy();
+    });
+  });
+});
+
+describe("getRelationOfInterface()", () => {
+  test("returns types and interfaces extending an interface", () => {
+    const schema = buildSchema(`
         interface Being {
           name(surname: Boolean): String
         }
@@ -483,11 +482,11 @@ describe("lib", () => {
         }
       `);
 
-      const interfaceType = schema.getType("Mammal");
+    const interfaceType = schema.getType("Mammal");
 
-      const relations = getRelationOfInterface(interfaceType, schema);
+    const relations = getRelationOfInterface(interfaceType, schema);
 
-      expect(relations).toMatchInlineSnapshot(`
+    expect(relations).toMatchInlineSnapshot(`
         {
           "interfaces": [
             "Canine",
@@ -497,11 +496,11 @@ describe("lib", () => {
           ],
         }
       `);
-    });
+  });
 
-    describe("getRelationOfUnion", () => {
-      test("returns unions using a type", () => {
-        const schema = buildSchema(`
+  describe("getRelationOfUnion", () => {
+    test("returns unions using a type", () => {
+      const schema = buildSchema(`
         type StudyItem {
           id: ID!
           subject: String!
@@ -524,23 +523,23 @@ describe("lib", () => {
         }
       `);
 
-        const compositeType = schema.getType("Meeting");
+      const compositeType = schema.getType("Meeting");
 
-        const relations = getRelationOfUnion(compositeType, schema);
+      const relations = getRelationOfUnion(compositeType, schema);
 
-        expect(relations).toMatchInlineSnapshot(`
+      expect(relations).toMatchInlineSnapshot(`
         {
           "unions": [
             "Task",
           ],
         }
       `);
-      });
     });
+  });
 
-    describe("getRelationOfImplementation", () => {
-      test("returns implementations compatible with type", () => {
-        const schema = buildSchema(`
+  describe("getRelationOfImplementation", () => {
+    test("returns implementations compatible with type", () => {
+      const schema = buildSchema(`
         interface Being {
           name(surname: Boolean): String
         }
@@ -573,11 +572,11 @@ describe("lib", () => {
         union Pet = Dog | Cat | Being
       `);
 
-        const compositeType = schema.getType("Being");
+      const compositeType = schema.getType("Being");
 
-        const relations = getRelationOfImplementation(compositeType, schema);
+      const relations = getRelationOfImplementation(compositeType, schema);
 
-        expect(relations).toMatchInlineSnapshot(`
+      expect(relations).toMatchInlineSnapshot(`
         {
           "interfaces": [
             "Canine",
@@ -591,11 +590,11 @@ describe("lib", () => {
           ],
         }
       `);
-      });
+    });
 
-      describe("getRelationOfReturn", () => {
-        test("returns queries, subscriptions and mutations using a type", () => {
-          const schema = buildSchema(`
+    describe("getRelationOfReturn", () => {
+      test("returns queries, subscriptions and mutations using a type", () => {
+        const schema = buildSchema(`
         type StudyItem {
           id: ID!
           subject: String!
@@ -616,17 +615,17 @@ describe("lib", () => {
         }
       `);
 
-          const compositeType = schema.getType("StudyItem");
+        const compositeType = schema.getType("StudyItem");
 
-          const relations = getRelationOfReturn(compositeType, schema);
+        const relations = getRelationOfReturn(compositeType, schema);
 
-          expect(relations).toMatchSnapshot();
-        });
+        expect(relations).toMatchSnapshot();
       });
+    });
 
-      describe("getRelationOfField", () => {
-        test("returns queries, subscriptions and mutations using a type", () => {
-          const schema = buildSchema(`
+    describe("getRelationOfField", () => {
+      test("returns queries, subscriptions and mutations using a type", () => {
+        const schema = buildSchema(`
         interface Record {
           id: String!
         }
@@ -651,12 +650,11 @@ describe("lib", () => {
         }
       `);
 
-          const compositeType = schema.getType("String");
+        const compositeType = schema.getType("String");
 
-          const relations = getRelationOfField(compositeType, schema);
+        const relations = getRelationOfField(compositeType, schema);
 
-          expect(relations).toMatchSnapshot();
-        });
+        expect(relations).toMatchSnapshot();
       });
     });
   });
